:truck: moved Component API objects to own folder
