+++ pincer/commands/__init__.py
@@ -5,11 +5,15 @@ from .arg_types import (
     CommandArg, Description, Choice, Choices, ChannelTypes, MaxValue, MinValue,
     Modifier
 )
-from .components import ComponentHandler, component, button
+from .components import (
+    ActionRow, Button, ButtonStyle, ComponentHandler, SelectMenu, component, button,
+    select_menu
+)
 from .commands import command, user_command, message_command, ChatCommandHandler
 
 __all__ = (
     "ChannelTypes", "ChatCommandHandler", "Choice", "Choices",
     "CommandArg", "Description", "MaxValue", "MinValue", "Modifier", "command",
-    "message_command", "user_command", "ComponentHandler", "component", "button"
+    "message_command", "user_command", "ComponentHandler", "component", "button",
+    "ActionRow", "Button", "ButtonStyle", "SelectMenu", "select_menu"
 )
